requirements.txt
streamlit==1.28.0
pandas==2.0.3
numpy==1.24.3
pillow==10.0.0
torch==2.0.1
torchvision>=0.15.0
transformers==4.35.0
scikit-learn==1.3.0
openpyxl==3.1.2
python-dotenv
